Use WeakIdConv by default and deprecate _id2ref version
